+++ lib/kitchen/directions/bake_appendix.rb
@@ -17,7 +17,9 @@ module Kitchen
 
         # Make a section with data-depth of X have a header level of X+1
         page.search('section').each do |section|
-          title = section.first!("[data-type='title']")
+          title = section.titles.first
+          next unless title.present?
+
           title.name = "h#{section['data-depth'].to_i + 1}"
         end
       end