+++ spec/directions/bake_appendix_spec.rb
@@ -7,13 +7,21 @@ RSpec.describe Kitchen::Directions::BakeAppendix do
     page_element(
       <<~HTML
         <div data-type="document-title">zzzzzzz</div>
-            <section data-depth="1">
-              <div data-type="title">hello</div>
-              <section data-depth="2">
-                <div data-type="title">world</div>
-              </section>
-            </section>
-        </div>
+        <section data-depth="1">
+          <div data-type="title">hello</div>
+          <section data-depth="2">
+            <div data-type="title">world</div>
+          </section>
+        </section>
+      HTML
+    )
+  end
+
+  let(:page_appendix_no_title) do
+    page_element(
+      <<~HTML
+        <div data-type="document-title">Hi</div>
+        <section data-depth="1"></section>
       HTML
     )
   end
@@ -35,6 +43,23 @@ RSpec.describe Kitchen::Directions::BakeAppendix do
               <h3 data-type="title">world</h3>
             </section>
           </section>
+        </div>
+      HTML
+    )
+  end
+
+  it 'does not explode if title not present in appendix section' do
+    described_class.v1(page: page_appendix_no_title, number: 3)
+    expect(page_appendix_no_title).to match_normalized_html(
+      <<~HTML
+        <div data-type="page">
+          <h1 data-type="document-title">
+            <span class="os-part-text">Appendix </span>
+            <span class="os-number">3</span>
+            <span class="os-divider"> </span>
+            <span class="os-text" data-type="" itemprop="">Hi</span>
+          </h1>
+          <section data-depth="1"></section>
         </div>
       HTML
     )